+++ src/main/java/HIS_E2/app_sanidad/model/Paciente.java
@@ -5,19 +5,19 @@ import org.springframework.data.mongodb.core.mapping.Document;
 @Document(collection = "paciente")
 public class Paciente extends Usuario {
 	
-	int numeroSS;
+	double numeroSS;
 
 
-	public Paciente(String dni, String nombre, String apellidos, String contrs, int numeroSS) {
+	public Paciente(String dni, String nombre, String apellidos, String contrs, double numeroSS) {
 		super(dni, nombre, apellidos, contrs);
 		this.numeroSS = numeroSS;
 	}
 
-	public int getNumeroSS() {
+	public double getNumeroSS() {
 		return numeroSS;
 	}
 
-	public void setNumeroSS(int numeroSS) {
+	public void setNumeroSS(double numeroSS) {
 		this.numeroSS = numeroSS;
 	}
